back/requirements.txt
flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0
mistralai==0.0.1
numpy==1.24.3
soundfile==0.12.1
SpeechRecognition==3.10.0
pydub==0.25.1
scipy==1.10.1
librosa==0.10.0
torch==2.0.1
torchaudio==2.0.2
onnxruntime==1.15.1
pymongo==4.4.1
dnspython==2.3.0